Added information about setting a static IP, DNS, and TLS certs
Included a diagram of the complicated setup for getting trusted TLS
certificates without ever connecting the HestiaPi to the internet.  This is a
complicated process, but if we document the complex process, then documenting
the simplier process in the future should be easy, as it will just be a matter
of providing the context and then referencing existing sections.

I have not gone through setting up the HestiaPi with TLS certificates yet,
so that section has been left as a TODO.